去掉服务器上的slua环境支持（支持没啥意义），补充基本玩家数据读写缓存的代码

+++ SluaExport/SluaManaged/Slua_Managed/LuaDLL.cs
@@ -92,7 +92,7 @@ namespace LuaInterface
 
         // Thread Funcs
         [DllImport(LUADLL, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int lua_tothread(IntPtr L, int index);
+        public static extern IntPtr lua_tothread(IntPtr L, int index);
         [DllImport(LUADLL, CallingConvention = CallingConvention.Cdecl)]
         public static extern void lua_xmove(IntPtr from, IntPtr to, int n);
